requirements.txt
amqp==2.1.0
anyjson==0.3.3
astroid==1.4.8
backport-collections==0.1
beautifulsoup4==4.5.1
billiard==3.5.0.2
boto==2.42.0
celery==4.0.2
coverage==4.2
cssselect==0.9.2
cssutils==1.0.1
diff-match-patch==20121119
dj-database-url==0.4.1
dj-static==0.0.6
django-absolute==0.3
django-admin-sortable2==0.6.4
django-admin-sortable==2.0.19
django-angular==0.8.4
django-appconf==1.0.2
django-appdata==0.1.5
-e git://github.com/kamilnowak/django-ckeditor.git#egg=django-ckeditor
django-classy-tags==0.8.0
django-cors-headers==1.2.1
django-emailit==0.2.2
django-filer==1.2.5
django-filter==0.15.2
django-formtools==1.0
django-haystack
django-mptt==0.8.6
django-parler==1.6.5
django-parler-rest==1.4.2
django-phonenumber-field==1.1.0
django-polymorphic==1.0.1
django-push-notifications==1.4.1
django-reversion==2.0.6
django-sekizai==0.10.0
django-simple-captcha==0.5.2
django-sizefield==0.9.1
django-sortedm2m==1.3.2
django-spurl==0.6.4
django-standard-form==1.1.1
django-storages==1.5.1
django-tablib==3.1.2
django-taggit==0.21.2
django-treebeard==4.0.1
Django==1.10.1
djangorestframework-gis==0.10.1
djangorestframework==3.4.7
djangorestframework-recursive==0.1.1
easy-thumbnails==2.3
elasticsearch==2.4.0
factory-boy==2.7.0
fake-factory==0.7.2
feedparser==5.2.1
flake8==3.0.4
future==0.15.2
gunicorn==19.6.0
jira==1.0.7
jsonfield==1.0.3
kombu==4.0.2
logilab-common==1.2.2
lxml==3.6.4
Markdown==2.6.7
mccabe==0.5.2
micawber==0.3.3
newrelic==2.70.0.51
oauthlib==2.0.0
pep8==1.7.0
phonenumbers==7.7.1
Pillow==3.3.1
premailer==3.0.1
psycopg2
pyflakes==1.3.0
pylint==1.6.4
python-dateutil==2.5.3
python-slugify==1.2.1
pytz
PyYAML==3.12
requests-mock==1.1.0
requests-oauthlib==0.7.0
requests==2.11.1
selenium==3.0.1
six==1.10.0
sorl-thumbnail==12.3
South==1.0.2
static3==0.7.0
tablib==0.11.2
tlslite==0.4.9
Unidecode==0.4.19
urllib3==1.18
URLObject==2.4.2
xlrd==1.0.0
xlwt-future==0.8.0
xmltodict==0.10.2
YURL==0.13
python-memcached
pycld2
google-api-python-client
google-cloud-translate
openpyxl
redis
html2text
pdfkit
